requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
djangorestframework-simplejwt
psycopg2-binary
gunicorn
python-dotenv
requests
pillow
drf_yasg
matplotlib
seaborn
pandas
numpy
